LLM/setup/requirements.txt
transformers
torch
numpy
pandas
rouge-score
bert-score
nltk
langchain
huggingface-hub
requests
PyPDF2
cleantext
tqdm
PyMuPDF
ads
pyarrow
scikit-learn
scipy
accelerate
sentencepiece
protobuf
pytest
marker-pdf[full] 
python-dotenv   
